third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-property-on-console-proto.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function testFunction()
{
    var amountOfProperties = 0;
    for (var p in {})
        ++amountOfProperties;
    console.__proto__.debug = 239;
    for (var p in {})
        --amountOfProperties;
    return amountOfProperties;
}

function test()
{
    InspectorTest.sendCommand("Runtime.evaluate", { "expression": "testFunction()" }, dumpResult);

    function dumpResult(result)
    {
        result.id = 0;
        InspectorTest.logObject(result);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">
Tests that property defined on console.__proto__ doesn't observable on other Objects.
</body>
</html>
