third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-property-on-console-proto-expected.txt

Tests that property defined on console.__proto__ doesn't observable on other Objects.
{
    id : <number>
    result : {
        result : {
            type : number
            value : 0
        }
    }
}
